Attempt to use desired theme, fall back to default
If the requirements.txt file is used with pip then the desired
theme will be available and our conditional theme choice will
not be necessary, but some users may overlook that setup step
and may try to build the docs without all required deps present.

This graceful fallback functionality is intended to work around
those cases.

refs rsyslog/rsyslog-doc#533