release build script: "wrap" reference Docker img
Instead of attempting to control sphinx-build directly this script
now attempts to use the reference Docker image to build a stable
release of the docs.

To work around issues with passing the required values via
environment variables a new "tool" entry was created to allow for
optional building of the docs using the 'better' theme.