Update tls.rst
Propose to update the client documentation to include the certificate. I was working on getting client rsyslog set up for over a week and not until I did detailed GnuTLS debugging on the destination I found that it would not accept my TLS syslogs unless I also sent the cert along with the client.